requirements.txt
asgiref==3.5.2
Django==3.2.14
django-extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.2.1
sqlparse==0.4.2
